requirements.txt
datasetsforecast
statsforecast
utilsforecast
seaborn
neuralforecast
prophet
transformers==4.40.1 
accelerate
scikit-learn
chronos-forecasting
lightgbm
mlforecast